+++ tests/timeseries/test_resampling.py
@@ -93,7 +93,7 @@ async def test_resampler_config_len_ok(
 ) -> None:
     """Test checks on the resampling buffer."""
     config = ResamplerConfig(
-        resampling_period_s=1.0,
+        resampling_period=timedelta(seconds=1.0),
         initial_buffer_len=init_len,
     )
     assert config.initial_buffer_len == init_len
@@ -110,7 +110,7 @@ async def test_resampler_config_len_warn(
 ) -> None:
     """Test checks on the resampling buffer."""
     config = ResamplerConfig(
-        resampling_period_s=1.0,
+        resampling_period=timedelta(seconds=1.0),
         initial_buffer_len=init_len,
     )
     assert config.initial_buffer_len == init_len
@@ -133,7 +133,7 @@ async def test_resampler_config_len_error(init_len: int) -> None:
     """Test checks on the resampling buffer."""
     with pytest.raises(ValueError):
         _ = ResamplerConfig(
-            resampling_period_s=1.0,
+            resampling_period=timedelta(seconds=1.0),
             initial_buffer_len=init_len,
         )
 
@@ -144,7 +144,7 @@ async def test_helper_buffer_too_big(
 ) -> None:
     """Test checks on the resampling buffer."""
     config = ResamplerConfig(
-        resampling_period_s=DEFAULT_BUFFER_LEN_MAX + 1,
+        resampling_period=timedelta(seconds=DEFAULT_BUFFER_LEN_MAX + 1),
         max_data_age_in_periods=1,
     )
     helper = _ResamplingHelper("test", config)
@@ -181,7 +181,7 @@ async def test_resampling_window_size_is_constant(
         spec=ResamplingFunction, return_value=expected_resampled_value
     )
     config = ResamplerConfig(
-        resampling_period_s=resampling_period_s,
+        resampling_period=timedelta(seconds=resampling_period_s),
         max_data_age_in_periods=1.0,
         resampling_function=resampling_fun_mock,
         initial_buffer_len=4,
@@ -265,7 +265,7 @@ async def test_timer_errors_are_logged(
         spec=ResamplingFunction, return_value=expected_resampled_value
     )
     config = ResamplerConfig(
-        resampling_period_s=resampling_period_s,
+        resampling_period=timedelta(seconds=resampling_period_s),
         max_data_age_in_periods=2.0,
         resampling_function=resampling_fun_mock,
         initial_buffer_len=4,
@@ -342,9 +342,9 @@ async def test_timer_errors_are_logged(
         "frequenz.sdk.timeseries._resampling",
         logging.WARNING,
         "The resampling task woke up too late. Resampling should have started at "
-        "1970-01-01 00:00:04+00:00, but it started at "
-        "1970-01-01 00:00:04.399800+00:00 (0.3998 seconds difference; resampling "
-        "period is 2 seconds)",
+        "1970-01-01 00:00:04+00:00, but it started at 1970-01-01 "
+        "00:00:04.399800+00:00 (tolerance: 0:00:00.200000, difference: "
+        "0:00:00.399800; resampling period: 0:00:02)",
     ) in _filter_logs(caplog.record_tuples, logger_level=logging.WARNING)
     sink_mock.reset_mock()
     resampling_fun_mock.reset_mock()
@@ -363,7 +363,7 @@ async def test_future_samples_not_included(
         spec=ResamplingFunction, return_value=expected_resampled_value
     )
     config = ResamplerConfig(
-        resampling_period_s=resampling_period_s,
+        resampling_period=timedelta(seconds=resampling_period_s),
         max_data_age_in_periods=2.0,
         resampling_function=resampling_fun_mock,
         initial_buffer_len=4,
@@ -448,7 +448,7 @@ async def test_resampling_with_one_window(
         spec=ResamplingFunction, return_value=expected_resampled_value
     )
     config = ResamplerConfig(
-        resampling_period_s=resampling_period_s,
+        resampling_period=timedelta(seconds=resampling_period_s),
         max_data_age_in_periods=1.0,
         resampling_function=resampling_fun_mock,
         initial_buffer_len=4,
@@ -557,7 +557,7 @@ async def test_resampling_with_one_and_a_half_windows(  # pylint: disable=too-ma
         spec=ResamplingFunction, return_value=expected_resampled_value
     )
     config = ResamplerConfig(
-        resampling_period_s=resampling_period_s,
+        resampling_period=timedelta(seconds=resampling_period_s),
         max_data_age_in_periods=1.5,
         resampling_function=resampling_fun_mock,
         initial_buffer_len=7,
@@ -714,7 +714,7 @@ async def test_resampling_with_two_windows(  # pylint: disable=too-many-statemen
         spec=ResamplingFunction, return_value=expected_resampled_value
     )
     config = ResamplerConfig(
-        resampling_period_s=resampling_period_s,
+        resampling_period=timedelta(seconds=resampling_period_s),
         max_data_age_in_periods=2.0,
         resampling_function=resampling_fun_mock,
         initial_buffer_len=16,
@@ -867,7 +867,7 @@ async def test_receiving_stopped_resampling_error(
         spec=ResamplingFunction, return_value=expected_resampled_value
     )
     config = ResamplerConfig(
-        resampling_period_s=resampling_period_s,
+        resampling_period=timedelta(seconds=resampling_period_s),
         max_data_age_in_periods=2.0,
         resampling_function=resampling_fun_mock,
     )
@@ -928,7 +928,7 @@ async def test_receiving_resampling_error(fake_time: time_machine.Coordinates) -
     )
     resampler = Resampler(
         ResamplerConfig(
-            resampling_period_s=resampling_period_s,
+            resampling_period=timedelta(seconds=resampling_period_s),
             max_data_age_in_periods=2.0,
             resampling_function=resampling_fun_mock,
         )