+++ RELEASE_NOTES.md
@@ -18,6 +18,7 @@
   Notice that the parameter `sampling_period` has been renamed to `input_sampling_period`
   to better distinguish it from the sampling period parameter in the resampler.
 * The serialization feature for the ringbuffer was made more flexible. The `dump` and `load` methods can now work directly with a ringbuffer instance.
+* The `ResamplerConfig` now takes the resampling period as a `timedelta`. The configuration was renamed from `resampling_period_s` to `resampling_period` accordingly.
 
 ## New Features